Upgrading the NTP timestamp
The NTP timestamp will now use a randomly generated value to generate a
cryptographic image in the insignificant bits of the fractional portion
of the timestamp. The image is based on the hashing of the random value,
the seconds, and the significant portion of the fractional part. It can
be used to determine whether the server instance generated a timestamp.

Tests are sorely needed, as the algorithm is a little complex and easy
to get wrong (coming shortly).