Adding test client
Test client sends 3 packets: (1) one with bad version number, (2) one
with invalid length, and then (3) a valid packet. Only a single response
is expected (from 3), and then that packet is replayed, but no response
is expected.

There is still an issue with higher number of rounds, where the test
fails on ocassion. Need to investigate.